requirements.txt
boto3
flask 
opencv-python
Pillow
flask-uploads